# Implement shape-inference caching for codegen LazyNativeFunctions