@wconstab 2022-01-25
Oops, didn't notice this is failing 3 tests:
[  FAILED  ] 3 tests, listed below:
[  FAILED  ] AtenLtcTsTensorTest.TestClampMax
[  FAILED  ] AtenLtcTsTensorTest.TestCelu
[  FAILED  ] AtenLtcTsTensorTest.TestCeluInPlace

not ready to merge
cc @desertfire